requirements.txt
absl-py==1.4.0
astunparse==1.6.3
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.3
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
dpcpp-cpp-rt==2023.0.0
editdistpy==0.1.3
flatbuffers==2.0.7
flit_core==3.6.0
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
ipdb==0.13.11
ipython==7.34.0
jacinle==1.0.2
jedi==0.18.2
joblib==1.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lark==1.1.3
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl==2023.1.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nltk==3.6.5
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
parso==0.8.3
peewee==3.14.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgconfig==1.5.5
plotly==5.13.1
plyfile==0.7.4
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
shapely==2.0.1
six==1.16.0
sklearn==0.0.post1
smplx==0.1.28
symspellpy==6.7.7
tabulate==0.9.0
tbb==2021.9.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
